Revert ":sparkles: Add New landpage of vulpes"
