--- components/VulpesLand/DataValuesSession/styles.module.css
@@ -1,104 +1,0 @@
-.container {
-    width: 86rem;
-    height: 100%;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    position: relative;
-    
-    letter-spacing: -.05rem;
-}
-
-.container__inside {
-    height: 100%;
-    width: 100%;
-    padding-top: 9rem;
-}
-
-.title {
-    color: white;
-    font-size: clamp(1.4rem, 3vw, 2rem);
-    font-weight: 500;
-    margin-bottom: 1.8rem;
-    z-index: 100;
-}
-
-.line {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    align-items: flex-start;
-    gap: 6rem;
-    margin-bottom: 6rem;
-    z-index: 100;
-}
-
-.line__info {
-    flex: 1;
-    font-size: clamp(2.2rem, 3vw ,3rem);
-    display: flex;
-    flex-direction: row;
-    gap: .3rem;
-}
-
-.line__value {
-    font-weight: 500;
-    color: var(--text-blue-color);
-}
-
-.line__unit {
-    color: white;
-}
-
-.break__line {
-    width: 80%;
-    margin-block: 0.4rem;
-    border-bottom: 1px solid #eaeaea;
-}
-
-.column__value {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-}
-
-.line__detail {
-    color: white;
-    width: fit-content;
-    max-width: 72%;
-}
-
-
-@media (max-width: 1400px) {
-    .container {
-        width: 80vw;
-    }
-
-    .line__detail {
-        max-width: none;
-    }
-}
-
-@media (max-width: 1050px) {
-    .line {
-        flex-direction: column;
-        align-items: center;
-        gap: 1.6rem;
-    }
-
-    .line__detail {
-        max-width: 50%;
-        text-align: center;
-    }
-
-    .break__line {
-        width: 10rem;
-    }
-
-    .line__info {
-        font-size: clamp(1.6rem, 3vw, 2.2rem);
-    }
-}
