--- components/VulpesLand/HeroSession/styles.module.css
@@ -1,88 +1,0 @@
-.container {
-    height: 40rem;
-    width: 86rem;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    position: relative;
-    
-    letter-spacing: -.05rem;
-}
-
-.intro {
-    margin-top: 14rem;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-}
-
-
-.gray {
-    color: #A2A2A2;
-    font-size: 2rem;
-}
-
-.title {
-    margin: 0;
-    font-size: 4rem;
-    letter-spacing: -.2rem;
-}
-
-.languages {
-    color: #42D3FF;
-}
-
-.paragr {
-    margin-bottom: 4rem;
-    font-size: 2rem;
-    width: 30rem;
-}
-
-.know__more__session {
-    display: flex;
-    flex-direction: row;
-    gap: 1rem;
-}
-
-.hand__arrow {
-    animation: jump 1.6s ease-in-out infinite;
-}
-
-.group__vulpes {
-    top: 5rem;
-    right: 6rem;
-    position: absolute;
-}
-
-.retangulos {
-    top: 0;
-    right: 0;
-    position: absolute;
-}
-
-
-@keyframes jump {
-	0% {
-		transform: translatey(.3rem);
-	}
-	50% {
-		transform: translatey(-.3rem);
-	}
-	100% {
-		transform: translatey(.3rem);
-	}
-}
-@keyframes float {
-	0% {
-		transform: translatey(0px);
-	}
-	50% {
-		transform: translatey(-1.6rem);
-	}
-	100% {
-		transform: translatey(0px);
-	}
-}
