--- components/VulpesLand/LandFooterSession/index.tsx
@@ -1,11 +1,0 @@
-import './styles.module.css';
-import styles from './styles.module.css';
-
-const LandFooterSession = () => {
-  return (
-    <div className={styles.container} id='header'>
-    </div>
-  );
-};
-
-export default LandFooterSession;
