--- components/VulpesLand/MethodologySession/index.tsx
@@ -1,12 +1,0 @@
-import './styles.module.css';
-import styles from './styles.module.css';
-
-const MethodologySession = () => {
-  return (
-    <div className={styles.container} id='header'>
-      
-    </div>
-  );
-};
-
-export default MethodologySession;
