--- components/VulpesLand/MethodologySession/styles.module.css
@@ -1,13 +1,0 @@
-.container {
-    width: 86rem;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    position: relative;
-    height: 100%;
-    
-    letter-spacing: -.05rem;
-}
