--- components/VulpesLand/NavSession/styles.module.css
@@ -1,68 +1,0 @@
-.container {
-    height: 40rem;
-    width: 86rem;
-    padding: 0;
-    margin: 0;
-    justify-content: flex-start;
-    align-items: center;
-    position: relative;
-    
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    
-    letter-spacing: -.05rem;
-}
-
-.logo {
-    font-size: 6rem;
-    height: 3rem;
-}
-
-.sessions {
-    height: 100%;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 1rem;
-}
-
-.session {
-    color: #777777;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-
-    height: 100%;
-    margin-bottom: -3px;
-
-    -webkit-transition: color .7s ease-out;
-    -moz-transition: color .7s ease-out;
-    -o-transition: color .7s ease-out;
-    transition: color .7s ease-out;
-}
-
-.session:hover {
-    color: black;
-    cursor: pointer;
-    padding-top: 0.2rem;
-    border-bottom: .2rem var(--text-red-color) solid;
-}
-
-.button {
-    background-color: transparent;
-    color: var(--text-red-color);
-    border: .1rem var(--text-red-color) solid;
-    font-weight: 600;
-    border-radius: .4rem;
-    padding: .2rem 1rem;
-    -webkit-transition: color .7s ease-out;
-    -moz-transition: color .7s ease-out;
-    -o-transition: color .7s ease-out;
-    transition: color .7s ease-out;
-}
-
-.button:hover {
-    background-color: var(--text-red-color);
-    color: #fefefe;
-}
