+++ package.json
@@ -15,7 +15,7 @@
     "next": "12.2.2",
     "react": "18.2.0",
     "react-bootstrap": "^2.4.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "18.2.0",
     "react-icons": "^4.8.0",
     "uuid": "^9.0.0"
   },
