--- public/LANDPAGE/WavesBottomSvg/styles.module.css
@@ -1,5 +1,0 @@
-.top__wave {
-    position: absolute;
-    bottom: -19rem;
-    z-index: 10;
-}
