--- public/LANDPAGE/WavesTopSvg/index.tsx
@@ -1,22 +1,0 @@
-import * as React from 'react'
-import './styles.module.css';
-import styles from './styles.module.css';
-
-type Colors = {
-  color__1: string,
-  color__2: string
-}
-
-const WavesTopSvg = ({ ...props }: Colors) => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1440 320" className={styles.top__wave}>
-    {/* <path fill={props.color__2} fillOpacity={1} d="M0,64L40,74.7C80,85,160,107,240,106.7C320,107,400,85,480,74.7C560,64,640,64,720,80C800,96,880,128,960,117.3C1040,107,1120,53,1200,32C1280,11,1360,21,1400,26.7L1440,32L1440,320L1400,320C1360,320,1280,320,1200,320C1120,320,1040,320,960,320C880,320,800,320,720,320C640,320,560,320,480,320C400,320,320,320,240,320C160,320,80,320,40,320L0,320Z"></path> */}
-    <path fill={props.color__2} fillOpacity={1} d="M0,256L80,261.3C160,267,320,277,480,256C640,235,800,181,960,170.7C1120,160,1280,192,1360,208L1440,224L1440,320L1360,320C1280,320,1120,320,960,320C800,320,640,320,480,320C320,320,160,320,80,320L0,320Z"></path>
-    <path fill={props.color__1} fillOpacity={1} d="M0,320L80,282.7C160,245,320,171,480,160C640,149,800,203,960,229.3C1120,256,1280,256,1360,256L1440,256L1440,320L1360,320C1280,320,1120,320,960,320C800,320,640,320,480,320C320,320,160,320,80,320L0,320Z"></path>
-
-    {/* <path fill={props.color__2} fillOpacity={1} d="M0,96L40,96C80,96,160,96,240,128C320,160,400,224,480,218.7C560,213,640,139,720,122.7C800,107,880,149,960,186.7C1040,224,1120,256,1200,234.7C1280,213,1360,139,1400,101.3L1440,64L1440,320L1400,320C1360,320,1280,320,1200,320C1120,320,1040,320,960,320C880,320,800,320,720,320C640,320,560,320,480,320C400,320,320,320,240,320C160,320,80,320,40,320L0,320Z"></path> */}
-    {/* <path fill={props.color__2} fillOpacity={1} d="M0,256L80,240C160,224,320,192,480,202.7C640,213,800,267,960,282.7C1120,299,1280,277,1360,266.7L1440,256L1440,320L1360,320C1280,320,1120,320,960,320C800,320,640,320,480,320C320,320,160,320,80,320L0,320Z"></path> */}
-
-  </svg>
-);
-
-export default WavesTopSvg;
