--- public/LANDPAGE/WavesTopSvg/styles.module.css
@@ -1,5 +1,0 @@
-.top__wave {
-    position: absolute;
-    top: -19rem;
-    z-index: 10;
-}
