--- public/LogoVulpes/index.tsx
@@ -1,58 +1,0 @@
-import * as React from 'react'
-import './styles.module.css';
-import styles from './styles.module.css';
-
-const SvgVulpesLogo = () => (
-  <svg viewBox='0 0 1161 394' fill='none' className={styles.svg__logo}>
-    <mask id='mask0_526_51' style={{maskType: 'luminance'}} maskUnits='userSpaceOnUse' x='1' y='1' width='330' height='393'>
-    <path d='M330.236 1.00684H1.00024V393.325H330.236V1.00684Z' fill='white'/>
-    </mask>
-    <g mask='url(#mask0_526_51)'>
-    <mask id='mask1_526_51' style={{maskType: 'luminance'}} maskUnits='userSpaceOnUse' x='1' y='1' width='330' height='393'>
-    <path d='M330.236 1.00684H1.00024V393.325H330.236V1.00684Z' fill='white'/>
-    </mask>
-    <g mask='url(#mask1_526_51)'>
-    <path d='M330.236 1L298.78 89.5394L233.087 52.5224L233.478 51.3924L330.236 1Z' className={styles.color__1}/>
-    <path d='M298.78 89.5396L298.834 144.171L178.168 276.282L152.273 286.795C152.256 286.774 152.242 286.751 152.232 286.726L298.793 89.5396' className={styles.color__2}/>
-    <path d='M298.78 89.5394L152.218 286.726C152.215 286.71 152.215 286.694 152.218 286.678L187.886 183.284L225.191 170.593C226.437 165.434 227.682 160.279 228.923 155.129L226.787 141.802L205.391 132.665L233.067 52.5361L298.78 89.5531' className={styles.color__1}/>
-    <path d='M297.924 227.841C298.531 233.331 298.835 238.851 298.834 244.374C298.884 271.809 291.316 298.719 276.974 322.106L152.328 286.904C152.306 286.87 152.287 286.833 152.273 286.795L178.195 276.302L297.944 227.841' className={styles.color__1}/>
-    <path d='M152.307 286.904L276.953 322.106C260.295 349.292 235.377 370.437 205.843 382.449L152.307 286.904Z' className={styles.color__1}/>
-    <path d='M226.787 141.774L228.923 155.102H108.949L104.408 89.5396L205.391 132.638L226.787 141.774Z' className={styles.color__1}/>
-    <path d='M228.923 155.102C227.677 160.261 226.433 165.416 225.191 170.566L187.886 183.256L109.216 210.021L85.4851 295.054L73.9109 372.444C41.273 353.049 17.2526 321.946 6.73926 285.466L65.0487 182.373L108.949 155.088L228.923 155.102Z' className={styles.color__2}/>
-    <path d='M152.307 286.904L205.864 382.45C184.585 391.049 161.608 394.623 138.722 392.895C115.836 391.167 93.6575 384.182 73.9109 372.485C73.9109 372.485 151.588 289.205 152.123 286.835L152.307 286.904Z' className={styles.color__1}/>
-    <path d='M109.216 210.021L152.184 286.678L152.123 286.849C151.602 289.205 73.9109 372.457 73.9109 372.457L85.4851 295.068L109.216 210.021Z' className={styles.color__1}/>
-    <path d='M104.408 89.5396L108.949 155.102L65.0488 182.387L11.2732 176.908C11.5129 176.285 11.7663 175.661 12.0197 175.038L17.389 163.395C31.7758 135.44 54.6092 112.731 82.643 98.4976C87.5976 95.9781 92.6921 93.7441 97.9018 91.8065C100.045 90.9983 102.216 90.2518 104.408 89.5464' className={styles.color__1}/>
-    <path d='M97.9018 51.3926V91.7996C92.692 93.7395 87.5974 95.9758 82.643 98.4976L78.5818 84.5195L97.9018 51.3926Z' className={styles.color__1}/>
-    <path d='M78.5816 84.5126L82.6429 98.4976C54.6173 112.728 31.7893 135.429 17.4026 163.375L68.9524 51.3994L78.5816 84.5126Z' className={styles.color__1}/>
-    <path d='M1.1508 238.21C1.04807 235.957 1.00013 233.683 1.00013 231.409C0.975339 212.774 4.46041 194.301 11.2731 176.956L65.0488 182.434L6.73931 285.466C2.9176 272.103 0.985779 258.272 1.00013 244.374C1.00013 242.292 1.04123 240.219 1.12341 238.155L1.1508 238.21Z' className={styles.color__1}/>
-    </g>
-    </g>
-    <path d='M12.0196 175.059C11.7662 175.682 11.5128 176.305 11.2731 176.928C4.46041 194.274 0.975339 212.747 1.00013 231.382C1.00013 233.656 1.04807 235.929 1.1508 238.183C1.1508 238.429 1.1508 238.669 1.1508 238.909' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M104.408 89.5396C102.216 90.2244 100.045 90.9915 97.9016 91.7996C92.6918 93.7395 87.5972 95.9758 82.6428 98.4976C54.6172 112.728 31.7892 135.429 17.4026 163.375C15.453 167.169 13.6632 171.05 12.0332 175.018' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M82.6429 98.4908L78.5816 84.5127L68.9524 51.3926L17.4026 163.368L12.0332 175.031L12.0195 175.059' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M97.9018 91.7996V51.3926L78.5818 84.5127' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M205.864 382.45L152.307 286.904C152.285 286.87 152.267 286.833 152.253 286.795C152.235 286.774 152.221 286.751 152.212 286.726C152.208 286.71 152.208 286.694 152.212 286.678L109.216 210.021L187.886 183.257L225.191 170.566C226.437 165.407 227.681 160.252 228.923 155.102L226.787 141.774L205.391 132.638L104.408 89.5396L108.949 155.102' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M152.123 286.836C151.602 289.191 73.9109 372.444 73.9109 372.444L85.4851 295.054L109.216 210' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M152.123 286.836L152.253 286.781L178.175 276.289L297.924 227.828C298.532 233.322 298.836 238.846 298.835 244.374C298.835 283.869 283.145 321.747 255.218 349.674C227.29 377.602 189.413 393.291 149.917 393.291C110.422 393.291 72.5445 377.602 44.6171 349.674C16.6897 321.747 1.00024 283.869 1.00024 244.374C1.00024 242.292 1.04134 240.219 1.12353 238.155' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M228.923 155.102H108.949L65.0487 182.387L6.73926 285.479' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M65.0488 182.387L11.2732 176.928' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M205.391 132.638L233.067 52.5224L233.457 51.3924L330.236 1L298.78 89.5394L298.834 144.171L178.168 276.282' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M276.98 322.106H276.974L152.307 286.904L152.123 286.849L152.184 286.678L187.893 183.284' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M232.807 52.3784L233.067 52.5222L298.78 89.5392L152.218 286.726L152.129 286.849' stroke='#FDF0D5' strokeWidth={1.0273} strokeLinejoin={'round'}/>
-    <path d='M410.344 53.244C394.629 59.22 382.456 68.516 373.824 81.132C365.192 93.5267 360.876 108.799 360.876 126.948C360.876 135.137 361.651 140.892 363.2 144.212C364.749 147.311 365.524 149.081 365.524 149.524C353.572 149.524 344.608 147.089 338.632 142.22C332.877 137.129 330 128.829 330 117.32C330 103.155 335.755 89.432 347.264 76.152C358.995 62.6507 373.935 51.8053 392.084 43.616C410.233 35.2053 428.383 31 446.532 31C449.631 31 454.057 31.2213 459.812 31.664L462.8 235.844L532.52 32.66H555.76L463.464 280H415.656L410.344 53.244Z' className={styles.color__1}/>
-    <path d='M565.058 281.992C554.434 281.992 545.802 278.672 539.162 272.032C532.744 265.392 529.534 255.432 529.534 242.152C529.534 236.619 530.42 229.425 532.19 220.572L554.766 114H602.574L578.67 226.88C578.006 229.536 577.674 232.303 577.674 235.18C577.674 244.919 580.994 249.788 587.634 249.788C593.832 249.788 599.144 247.243 603.57 242.152C607.997 236.84 611.538 229.647 614.194 220.572L636.77 114H684.578L660.674 226.88C659.789 230.2 659.346 233.741 659.346 237.504C659.346 241.931 660.342 245.14 662.334 247.132C664.548 248.903 668.089 249.788 672.958 249.788C676.5 249.788 679.045 249.567 680.594 249.124C682.365 248.46 683.472 248.017 683.914 247.796C683.029 259.084 679.156 267.605 672.294 273.36C665.433 279.115 657.022 281.992 647.062 281.992C637.766 281.992 630.241 279.225 624.486 273.692C618.953 268.159 615.633 260.412 614.526 250.452C607.001 262.625 599.033 270.925 590.622 275.352C582.433 279.779 573.912 281.992 565.058 281.992Z' className={styles.color__1}/>
-    <path d='M732.061 281.992C721.437 281.992 712.805 278.672 706.165 272.032C699.747 265.171 696.537 254.989 696.537 241.488C696.537 235.955 697.423 228.761 699.193 219.908L734.717 53.576L783.853 46.936L745.673 226.216C744.788 229.536 744.345 232.745 744.345 235.844C744.345 244.697 748.883 249.124 757.957 249.124C761.941 249.124 765.483 248.46 768.581 247.132C767.696 258.199 763.712 266.831 756.629 273.028C749.768 279.004 741.579 281.992 732.061 281.992Z' className={styles.color__1}/>
-    <path d='M793.81 101.716H841.618L836.306 126.616C847.373 117.098 860.321 112.34 875.15 112.34C887.545 112.34 897.394 116.434 904.698 124.624C912.002 132.813 915.654 146.204 915.654 164.796C915.654 182.281 913.109 199.988 908.018 217.916C902.927 235.622 894.295 250.784 882.122 263.4C869.949 275.794 853.902 281.992 833.982 281.992C819.817 281.992 810.41 278.008 805.762 270.04L788.498 351.048L738.366 363L793.81 101.716ZM823.69 253.108C834.314 253.108 843.057 248.128 849.918 238.168C857.001 228.208 862.091 216.366 865.19 202.644C868.51 188.7 870.17 175.752 870.17 163.8C870.17 145.429 864.637 136.244 853.57 136.244C849.586 136.244 845.491 137.682 841.286 140.56C837.302 143.437 833.761 147.421 830.662 152.512L811.074 245.472C812.845 250.562 817.05 253.108 823.69 253.108Z' className={styles.color__1}/>
-    <path d='M976.08 281.992C959.701 281.992 946.974 277.787 937.9 269.376C928.825 260.744 924.288 247.353 924.288 229.204C924.288 213.932 927.276 197.111 933.252 178.74C939.228 160.369 948.966 144.544 962.468 131.264C975.969 117.763 993.122 111.012 1013.93 111.012C1038.27 111.012 1050.45 121.636 1050.45 142.884C1050.45 155.279 1046.91 166.677 1039.82 177.08C1032.74 187.483 1023.33 195.893 1011.6 202.312C999.873 208.509 987.368 212.051 974.088 212.936C973.645 219.576 973.424 224.003 973.424 226.216C973.424 237.061 975.305 244.476 979.068 248.46C982.83 252.223 988.917 254.104 997.328 254.104C1009.28 254.104 1019.46 251.337 1027.87 245.804C1036.5 240.271 1045.91 231.86 1056.09 220.572H1067.38C1042.81 261.519 1012.38 281.992 976.08 281.992ZM976.744 197C984.933 196.557 992.68 193.68 999.984 188.368C1007.51 183.056 1013.48 176.305 1017.91 168.116C1022.56 159.927 1024.88 151.295 1024.88 142.22C1024.88 133.145 1022.12 128.608 1016.58 128.608C1008.62 128.608 1000.76 135.58 993.012 149.524C985.486 163.468 980.064 179.293 976.744 197Z' className={styles.color__1}/>
-    <path d='M1093.87 281.992C1083.25 281.992 1074.28 280 1066.98 276.016C1059.67 271.811 1054.25 266.499 1050.71 260.08C1047.17 253.44 1045.4 246.579 1045.4 239.496C1045.4 232.192 1047.17 225.884 1050.71 220.572C1054.03 215.039 1058.12 211.055 1062.99 208.62C1071.63 193.127 1079.15 177.523 1085.57 161.808C1091.99 145.872 1098.07 128.719 1103.83 110.348L1152.97 103.708C1154.07 132.039 1155.95 162.693 1158.61 195.672C1159.72 208.952 1160.27 218.58 1160.27 224.556C1160.27 238.057 1156.84 249.124 1149.98 257.756C1143.12 266.388 1134.59 272.585 1124.41 276.348C1114.45 280.111 1104.27 281.992 1093.87 281.992ZM1083.25 255.764C1091.43 255.764 1098.3 253.329 1103.83 248.46C1109.36 243.591 1112.13 235.733 1112.13 224.888C1112.13 218.248 1111.47 208.952 1110.14 197C1107.92 171.768 1106.49 154.836 1105.82 146.204C1100.51 163.689 1091.55 184.605 1078.93 208.952C1084.02 211.608 1086.57 215.481 1086.57 220.572C1086.57 224.777 1085.13 228.54 1082.25 231.86C1079.59 235.18 1076.16 236.84 1071.96 236.84C1067.31 236.84 1064.32 235.401 1062.99 232.524C1062.99 240.492 1064.54 246.357 1067.64 250.12C1070.96 253.883 1076.16 255.764 1083.25 255.764Z' className={styles.color__1}/>
-    <path d='M422.858 392.514C421.79 392.514 421.256 391.083 421.256 388.221V332.281C421.256 329.462 421.79 328.052 422.858 328.052H430.035C431.06 328.052 431.573 329.462 431.573 332.281V388.221C431.573 391.083 431.06 392.514 430.035 392.514H422.858ZM422.602 318.12C421.534 318.12 421 316.967 421 314.66V306.394C421 304.13 421.534 302.998 422.602 302.998H430.227C431.295 302.998 431.829 304.13 431.829 306.394V314.66C431.829 316.967 431.295 318.12 430.227 318.12H422.602Z' className={styles.color__1}/>
-    <path d='M459.575 393.219C448.34 393.219 442.722 382.304 442.722 360.475C442.722 338.347 448.19 327.283 459.126 327.283C463.868 327.283 467.371 330.274 469.635 336.254C469.934 336.681 470.105 336.895 470.147 336.895C470.318 336.895 470.404 336.617 470.404 336.062V304.856C470.404 302.293 470.938 301.012 472.005 301.012H479.951C481.019 301.012 481.553 302.293 481.553 304.856V388.221C481.553 391.083 481.04 392.514 480.015 392.514H478.157C477.217 392.514 476.491 391.232 475.978 388.669C474.825 382.091 474.141 378.759 473.928 378.673C473.8 378.801 473.65 379.058 473.479 379.442C472.326 384.739 470.574 388.37 468.225 390.335C465.875 392.258 462.992 393.219 459.575 393.219ZM470.339 359.771C470.339 351.697 469.613 345.716 468.161 341.829C466.708 337.942 464.679 335.998 462.074 335.998C460.792 335.998 459.681 336.34 458.742 337.023C457.844 337.664 457.011 338.817 456.242 340.483C455.474 342.149 454.897 344.606 454.512 347.852C454.128 351.056 453.936 355.029 453.936 359.771C453.936 366.392 454.277 371.561 454.961 375.277C455.644 378.994 456.542 381.471 457.652 382.71C458.763 383.906 460.215 384.504 462.009 384.504C467.563 384.504 470.339 376.26 470.339 359.771Z' className={styles.color__1}/>
-    <path d='M496.803 392.514C495.735 392.514 495.201 391.083 495.201 388.221V332.281C495.201 329.462 495.735 328.052 496.803 328.052H503.98C505.005 328.052 505.518 329.462 505.518 332.281V388.221C505.518 391.083 505.005 392.514 503.98 392.514H496.803ZM496.547 318.12C495.479 318.12 494.945 316.967 494.945 314.66V306.394C494.945 304.13 495.479 302.998 496.547 302.998H504.172C505.24 302.998 505.774 304.13 505.774 306.394V314.66C505.774 316.967 505.24 318.12 504.172 318.12H496.547Z' className={styles.color__1}/>
-    <path d='M532.815 383.671C533.755 384.227 534.908 384.504 536.275 384.504C537.642 384.504 538.774 384.227 539.671 383.671C540.568 383.073 541.401 381.963 542.17 380.339C542.939 378.673 543.516 376.174 543.9 372.842C544.285 369.468 544.477 365.217 544.477 360.091C544.477 355.093 544.285 350.949 543.9 347.66C543.516 344.371 542.918 341.914 542.106 340.291C541.337 338.625 540.504 337.493 539.607 336.895C538.71 336.297 537.599 335.998 536.275 335.998C534.951 335.998 533.819 336.297 532.879 336.895C531.982 337.493 531.127 338.625 530.316 340.291C529.547 341.914 528.97 344.371 528.586 347.66C528.201 350.907 528.009 355.05 528.009 360.091C528.009 365.217 528.18 369.446 528.522 372.778C528.906 376.11 529.483 378.609 530.252 380.275C531.063 381.941 531.918 383.073 532.815 383.671ZM539.991 393.027C538.924 393.155 537.685 393.219 536.275 393.219C534.865 393.219 533.605 393.155 532.494 393.027C531.426 392.898 530.188 392.578 528.778 392.065C527.368 391.553 526.129 390.848 525.061 389.951C524.036 389.054 522.968 387.772 521.858 386.106C520.747 384.44 519.85 382.475 519.166 380.211C518.483 377.947 517.906 375.106 517.436 371.689C517.009 368.229 516.795 364.363 516.795 360.091C516.795 355.819 517.009 351.996 517.436 348.621C517.906 345.204 518.504 342.363 519.23 340.099C519.957 337.835 520.875 335.891 521.986 334.268C523.096 332.602 524.164 331.342 525.19 330.487C526.258 329.59 527.496 328.907 528.906 328.437C530.358 327.924 531.597 327.604 532.623 327.476C533.691 327.347 534.908 327.283 536.275 327.283C537.642 327.283 538.838 327.347 539.863 327.476C540.931 327.604 542.17 327.924 543.58 328.437C544.989 328.907 546.207 329.59 547.232 330.487C548.3 331.342 549.389 332.602 550.5 334.268C551.611 335.891 552.508 337.835 553.191 340.099C553.918 342.363 554.494 345.204 554.921 348.621C555.391 351.996 555.626 355.819 555.626 360.091C555.626 364.363 555.391 368.229 554.921 371.689C554.494 375.106 553.939 377.947 553.255 380.211C552.572 382.475 551.675 384.44 550.564 386.106C549.496 387.772 548.428 389.054 547.36 389.951C546.335 390.848 545.118 391.553 543.708 392.065C542.298 392.578 541.059 392.898 539.991 393.027Z' className={styles.color__1}/>
-    <path d='M569.083 392.514C568.613 392.514 568.228 392.13 567.929 391.361C567.63 390.549 567.481 389.481 567.481 388.157V332.73C567.481 331.363 567.609 330.252 567.865 329.398C568.164 328.501 568.549 328.052 569.018 328.052H571.261C572.329 328.052 573.098 329.334 573.568 331.897L574.721 340.547C574.849 341.06 575.02 341.316 575.234 341.316C575.362 341.316 575.49 341.039 575.618 340.483C577.797 331.683 581.877 327.283 587.857 327.283C593.325 327.283 597.127 331.342 599.263 339.458C599.391 339.8 599.519 339.971 599.647 339.971C599.904 339.971 600.053 339.8 600.096 339.458C601.164 335.443 602.83 332.41 605.094 330.359C607.401 328.309 609.75 327.283 612.142 327.283C613.808 327.283 615.261 327.476 616.5 327.86C617.739 328.245 618.956 329.013 620.152 330.167C621.391 331.278 622.395 332.794 623.164 334.716C623.933 336.639 624.552 339.223 625.022 342.47C625.492 345.716 625.727 349.54 625.727 353.94V388.157C625.727 391.062 625.193 392.514 624.125 392.514H616.692C615.624 392.514 615.09 391.062 615.09 388.157V352.786C615.09 346.293 614.663 341.872 613.808 339.522C612.954 337.173 611.523 335.998 609.515 335.998C607.337 335.998 605.521 337.514 604.069 340.547C602.659 343.538 601.954 348.301 601.954 354.837V388.157C601.954 389.481 601.805 390.549 601.506 391.361C601.207 392.13 600.822 392.514 600.352 392.514H592.983C591.873 392.514 591.317 391.062 591.317 388.157V353.619C591.317 346.912 590.869 342.299 589.972 339.778C589.075 337.258 587.665 335.998 585.743 335.998C584.034 335.998 582.581 336.916 581.385 338.753C580.232 340.547 579.399 342.876 578.886 345.738C578.416 348.557 578.182 351.825 578.182 355.541V388.157C578.182 389.481 578.032 390.549 577.733 391.361C577.434 392.13 577.049 392.514 576.58 392.514H569.083Z' className={styles.color__1}/>
-    <path d='M649.884 393.283C645.655 393.283 642.366 391.916 640.016 389.182C637.709 386.448 636.556 382.475 636.556 377.264C636.556 373.376 637.133 369.98 638.286 367.075C639.482 364.171 640.956 361.971 642.707 360.475C644.501 358.98 646.466 357.741 648.602 356.759C650.738 355.776 652.725 355.136 654.562 354.837C656.398 354.538 658 354.26 659.367 354.004C660.777 353.705 661.589 353.406 661.802 353.107V349.71C661.802 345.14 661.29 341.722 660.264 339.458C659.282 337.151 657.466 335.998 654.818 335.998C652.597 335.998 650.888 336.703 649.692 338.112C648.538 339.479 647.556 341.594 646.744 344.456C646.36 345.652 645.847 346.208 645.206 346.122C640.721 345.011 638.457 344.285 638.414 343.943C638.158 343.815 638.008 343.324 637.966 342.47C637.966 341.615 638.072 340.825 638.286 340.099C639.397 335.912 641.298 332.751 643.989 330.615C646.68 328.437 650.525 327.347 655.523 327.347C658.214 327.347 660.542 327.796 662.507 328.693C664.515 329.547 666.266 330.979 667.761 332.986C669.257 334.994 670.367 337.728 671.094 341.188C671.862 344.606 672.247 348.792 672.247 353.747V389.31C672.247 391.489 671.798 392.578 670.901 392.578H668.851C668.082 392.578 667.398 391.638 666.8 389.759C666.715 389.332 666.288 387.943 665.519 385.594C664.793 383.244 664.387 381.834 664.301 381.365C664.259 381.236 664.195 381.151 664.109 381.108C664.066 381.066 664.002 381.066 663.917 381.108C663.831 381.151 663.789 381.236 663.789 381.365C662.635 385.295 660.969 388.264 658.791 390.271C656.655 392.279 653.686 393.283 649.884 393.283ZM652.127 385.081C655.117 385.081 657.509 383.116 659.303 379.186C661.14 375.256 662.059 370.557 662.059 365.089C662.059 361.842 661.866 360.219 661.482 360.219C660.713 360.219 660.072 360.347 659.56 360.604C654.946 361.757 651.55 363.637 649.371 366.242C647.235 368.848 646.167 372.244 646.167 376.431C646.167 382.198 648.154 385.081 652.127 385.081Z' className={styles.color__1}/>
-    <path d='M696.468 393.219C693.521 393.219 690.637 392.514 687.818 391.104C685.041 389.695 682.606 387.73 680.513 385.209L684.293 377.84C688.181 382.283 692.111 384.504 696.084 384.504C698.519 384.504 700.355 383.671 701.594 382.005C702.876 380.297 703.517 378.225 703.517 375.79C703.517 373.44 702.854 371.518 701.53 370.023C700.249 368.485 697.921 366.541 694.546 364.192C692.538 362.782 690.915 361.565 689.676 360.539C688.48 359.472 687.305 358.169 686.152 356.631C684.998 355.093 684.144 353.363 683.589 351.441C683.076 349.475 682.82 347.211 682.82 344.648C682.82 339.65 684.208 335.507 686.985 332.217C689.761 328.928 693.392 327.283 697.878 327.283C703.303 327.283 707.959 329.441 711.847 333.755L707.874 341.38C704.841 338.048 701.786 336.382 698.711 336.382C696.703 336.382 695.144 337.109 694.033 338.561C692.922 339.971 692.367 341.743 692.367 343.879C692.367 346.272 693.029 348.301 694.354 349.967C695.678 351.59 697.686 353.256 700.377 354.965C702.299 356.204 703.922 357.4 705.247 358.553C706.614 359.706 707.917 361.116 709.155 362.782C710.437 364.405 711.398 366.306 712.039 368.485C712.68 370.664 713 373.12 713 375.854C713 381.236 711.548 385.487 708.643 388.605C705.781 391.681 701.722 393.219 696.468 393.219Z' className={styles.color__1}/>
-  </svg>
-);
-
-export default SvgVulpesLogo;
