--- public/LogoVulpes/styles.module.css
@@ -1,22 +1,0 @@
-.color__1, .color__2 {
-    fill: var(--text-dark-color);
-    -webkit-transition: fill .7s ease-out;
-    -moz-transition: fill .7s ease-out;
-    -o-transition: fill .7s ease-out;
-    transition: fill .7s ease-out;
-}
-
-.svg__logo {
-    height: 3rem;
-}
-
-.svg__logo:hover {
-    cursor: pointer;
-
-    .color__1 {
-        fill: #780000;
-    }
-    .color__2 {
-        fill: #E0A38D;
-    }
-}
