+++ styles/theme.css
@@ -4,7 +4,6 @@
   --text-blue-color: #42D3FF;
   --text-red-color: #C1121F;
   --black-background: #0f0f0f;
-  --marsala: #7B3838;
   --gradient-red: transparent linear-gradient(180deg, #C1121F 0%, #9F1D27 100%) 0% 0% no-repeat padding-box;
   --gradient-red-hover: transparent linear-gradient(180deg, #C1121F 70%, #9F1D27 100%) 0% 0% no-repeat padding-box;
   --gradient-blue: transparent linear-gradient(180deg, #11628D 0%, #003049 100%) 0% 0% no-repeat padding-box;
