--- views/VulpesLand/DataValuesView/styles.module.css
@@ -1,10 +1,0 @@
-.container {
-    height: 4rem;
-    width: 100vw;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    background-color: #fefefe;
-}
-
