--- views/VulpesLand/HeroView/styles.module.css
@@ -1,12 +1,0 @@
-.container {
-    height: 110vh;
-    width: 100vw;
-    padding: 0;
-    margin: 5rem 0 0 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    overflow-y: hidden;
-    background-color: #fefefe;
-}
-
