--- views/VulpesLand/LandFooterView/index.tsx
@@ -1,13 +1,0 @@
-import LandFooterSession from '../../../components/VulpesLand/LandFooterSession';
-import './styles.module.css';
-import styles from './styles.module.css';
-
-const LandFooterView = () => {
-  return (
-    <div className={styles.container} id='header'>
-      <LandFooterSession></LandFooterSession>
-    </div >
-  );
-};
-
-export default LandFooterView;
