--- views/VulpesLand/LandFooterView/styles.module.css
@@ -1,11 +1,0 @@
-.container {
-    height: 30rem;
-    width: 100vw;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    overflow-y: hidden;
-    background-color: var(--black-background);
-}
-
