--- views/VulpesLand/NavView/styles.module.css
@@ -1,16 +1,0 @@
-.container {
-    height: 5rem;
-    width: 100vw;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    overflow-y: hidden;
-    background-color: #fefefe;
-    border-block: .1rem var(--placeholder) solid;
-    position: fixed;
-    top: 0;
-    z-index: 999;
-}
-
